Remove instructions for `-trezor` from changelog
It took me a while to figure out why I couldn't get the Trezor implementation working. The only mention of the system is with instructions the changelog, so this just adds a small update to indicate the support is no longer available.

On Ubuntu it is also in the man page, but I'm not sure where the source of that is.